refactoring for file organization + gh cloning code
