+++ .gitignore
@@ -161,3 +161,4 @@ cython_debug/
 
 *.env
 *.env.local
+temporary_model_storage/
