+++ requirements.txt
@@ -4,3 +4,5 @@ yfinance==0.2.51
 psycopg2==2.9.10
 sqlalchemy==2.0.36
 docker==7.1.0
+redis==5.2.1
+huey==2.5.2
